adding ability to load fixtures from bundle different then the one CMFixture resides in
added CMFixture:setBundleForLoadingFixtures, which can be used to set
bundle in which CMFixture will look for fixtures. This is useful when
CMFixture is used as framework ( for example when using cocoapods with
'use_frameworks!'), since in that case the bundle from which fixtures are
loaded by default is the CMFixture framework bundle.